refactor(models): pull models out of responses dir
