+++ src/otf_api/models/book_class.py
