+++ src/otf_api/models/challenge_tracker_content.py
