+++ src/otf_api/models/favorite_studios.py
