+++ src/otf_api/models/member_detail.py
