+++ src/otf_api/models/out_of_studio_workout_history.py
