+++ src/otf_api/models/performance_summary_detail.py
