+++ src/otf_api/models/studio_detail.py
