+++ src/otf_api/models/telemetry_hr_history.py
