+++ src/otf_api/models/telemetry_max_hr.py
