+++ src/otf_api/models/total_classes.py
